requirements.txt
aniso8601==9.0.1
blinker==1.6.2
cachelib==0.10.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
cx-Oracle==8.3.0
Flask==2.2.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-Session==0.5.0
gunicorn==19.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
Werkzeug==2.3.6
